chore: Move twitter auth to account scoped
